+++ CHANGELOG.md
@@ -15,6 +15,8 @@ this project uses date-based 'snapshot' version identifiers.
 ### Changed
 
 - Expand default value of `binaryfiles`
+- Normalize Lua-only attribute, `\XeTeXcharclass`
+  allocation lines and user defined `\whatsit` nodes (may require `.tlg` update)
 
 ### Fixed